Move suite.util.build_git_url to repo_utils
