+++ teuthology/repo_utils.py
@@ -1,6 +1,7 @@
 import fcntl
 import logging
 import os
+import re
 import shutil
 import subprocess
 import time
@@ -34,6 +35,20 @@ def is_fresh(path):
     elif time.time() - os.stat(path).st_mtime < FRESHNESS_INTERVAL:
         return True
     return False
+
+
+def build_git_url(project, project_owner='ceph'):
+    """
+    Return the git URL to clone the project
+    """
+    if project == 'ceph-qa-suite':
+        base = config.get_ceph_qa_suite_git_url()
+    elif project == 'ceph':
+        base = config.get_ceph_git_url()
+    else:
+        base = 'https://github.com/{project_owner}/{project}'
+    url_templ = re.sub('\.git$', '', base)
+    return url_templ.format(project_owner=project_owner, project=project)
 
 
 def ls_remote(url, ref):
